Additional refactors
