+++ packages/excalidraw/renderer/renderElement.ts
@@ -27,7 +27,7 @@ import type {
   InteractiveCanvasRenderConfig,
 } from "../scene/types";
 import { distance, getFontString, isRTL } from "../utils";
-import { getCornerRadius, isRightAngle } from "../math";
+import { getCornerRadius } from "../math";
 import rough from "roughjs/bin/rough";
 import type {
   AppState,
@@ -907,7 +907,8 @@ export const renderElement = (
           (!element.angle ||
             // or check if angle is a right angle in which case we can still
             // disable smoothing without adversely affecting the result
-            isRightAngle(element.angle))
+            // We need less-than comparison because of FP artihmetic
+            Math.abs(Math.sin(2 * element.angle)) < 0.00001)
         ) {
           // Disabling smoothing makes output much sharper, especially for
           // text. Unless for non-right angles, where the aliasing is really
