+++ packages/math/ga/galines.ts
