+++ packages/utils/withinBounds.ts
@@ -11,12 +11,11 @@ import {
   isLinearElement,
   isTextElement,
 } from "../excalidraw/element/typeChecks";
-import { isValueInRange } from "../excalidraw/math";
 import type { Bounds } from "../excalidraw/element/bounds";
 import { getElementBounds } from "../excalidraw/element/bounds";
 import { arrayToMap } from "../excalidraw/utils";
 import type { LocalPoint } from "../math";
-import { point, pointRotateRads } from "../math";
+import { isValueInRange, point, pointRotateRads } from "../math";
 
 type Element = NonDeletedExcalidrawElement;
 type Elements = readonly NonDeletedExcalidrawElement[];
